+++ assets/features/library-content/library-speaker-admin.js
@@ -1,4 +1,4 @@
-/* TSOL Library Speaker profile editor */
+/* Member Library Speaker profile editor */
 
 (function($, wp, settings) {
     'use strict';